requirements.txt
openai>=1.30.0
python-dotenv>=1.0.1
tqdm>=4.66.0
Flask>=3.0.0
